very major commit after a long time

+++ src/main/resources/db/migration/mysql/V3__table_create.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS driver (
     id                           INTEGER      NOT NULL PRIMARY KEY AUTO_INCREMENT,
     user_id                      INTEGER      NOT NULL,
     ready                        VARCHAR(10)  NOT NULL,
-    lat                          DOUBLE(10,10) NOT NULL,
-    lng                          DOUBLE(10,10) NOT NULL,
+    lat                          DECIMAL(50,30) NOT NULL,
+    lng                          DECIMAL(50,30) NOT NULL,
     FOREIGN KEY (user_id) REFERENCES user (id)
     ON UPDATE CASCADE
     ON DELETE CASCADE
